refactor(core): Split planner run into briefing + coordinator modules
The plan tool file had accreted three distinct responsibilities — message
briefing, planner-run coordination, and the tool definition — and was
over 1.1k lines. Separate them into a clean dependency line:

- planner-briefing.ts: recent-message recall, prior-tool-observation
  reconstruction, and briefing formatting (getRecentMessages,
  getPriorToolObservations, buildPlannerBriefingContext,
  formatMessagesForBriefing).
- planner-run-coordinator.ts: PlannerRunCoordinator + its build/cleanup
  helpers and planToolSuspendSchema; imports briefing.
- plan-with-agent.tool.ts: just the Tool definition + thin handler;
  imports the coordinator. ~105 lines.

No file now exceeds ~520 lines and the import graph is acyclic
(tool → coordinator → briefing).

Drop the `__test*` re-export indirection: the helpers are now legitimate
module exports, so tests import them directly. The old single test file
splits to mirror the modules (planner-briefing.test.ts,
planner-run-coordinator.test.ts). Behaviour-preserving — all 205
orchestration tests pass.
